[#66] Set the new homepage and set block contents
We can't set block contents during site install because the
body field is not installed yet when dcamp_install() runs. Therefore,
I have updated README so it is now required to download the db
from Jenkins in order to install the site. This seems realistic to me
since now we have a production site so we need to work on database
updates from now on rather than installing a site.